API: Rename files appropriately.
