+++ as-ap-sofb/scripts/sirius-ioc-bo-ap-sofb.py
@@ -1,0 +1,10 @@
+#!/usr/local/bin/python-sirius -u
+
+import argparse as _argparse
+from as_ap_sofb import as_ap_sofb
+
+parser = _argparse.ArgumentParser(description="Run BO SOFB IOC.")
+parser.add_argument('-d', '--debug', action='store_true', default=False,
+                    help="Starts IOC in Debug Mode.")
+args = parser.parse_args()
+as_ap_sofb.run(acc='BO', debug=args.debug)
