+++ as-ap-sofb/scripts/sirius-ioc-si-ap-sofb.py
@@ -1,10 +1,10 @@
 #!/usr/local/bin/python-sirius -u
 
 import argparse as _argparse
-from si_ap_sofb import si_ap_sofb
+from as_ap_sofb import as_ap_sofb
 
 parser = _argparse.ArgumentParser(description="Run SI SOFB IOC.")
 parser.add_argument('-d', '--debug', action='store_true', default=False,
                     help="Starts IOC in Debug Mode.")
 args = parser.parse_args()
-si_ap_sofb.run(debug=args.debug)
+as_ap_sofb.run(acc='SI', debug=args.debug)
